feat: expand CrocSettings model and overhaul ReceiveScreen UI with animated state transitions and enhanced input features

+++ app/src/main/kotlin/com/henkenlink/crocdroid/domain/model/TransferHistory.kt
@@ -14,7 +14,7 @@ data class HistoryEntry(
     val timestamp: Long,
     val fileName: String,
     val fileSize: Long,
-    val fileCount: Int,
+    val fileCount: Long,
     val success: Boolean,
     val errorMessage: String? = null
 )
